Add files via upload
In this project work on utilizing LSTM (Long Short-Term Memory) for stock price prediction. LSTM is a type of recurrent neural network (RNN) that is widely used for sequence modeling and has shown great success in time-series forecasting. Through this project, I was able to apply LSTM to historical stock price data, and train the model to predict future prices with high accuracy. This approach has enormous potential to benefit the finance industry and help investors make informed decisions.